pyproject.toml
[project]
name = "classificador-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "llama-index>=0.13.3",
    "llama-index-llms-groq>=0.4.0",
    "llama-index-readers-file>=0.5.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]
